solver moves
